back/requirements.txt
Django
asgiref==3.3.2
django-cors-headers==3.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
djongo
pymongo==3.11.4
pytz==2021.1
setuptools==57.5.0
sqlparse==0.2.4
tzdata==2021.1
PyJWT==2.6.0
bcrypt
